Rename feature folder and namespaces from Item to Items
