+++ templates/MicroserviceTemplate.Aspire/src/MyService.Api/Common/Extensions/EndpointExtensions.cs
@@ -1,4 +1,4 @@
-using MyService.Api.Item.Endpoints;
+using MyService.Api.Items.Endpoints;
 
 namespace MyService.Api.Common.Extensions;
 
@@ -6,7 +6,7 @@ public static class EndpointExtensions
 {
     public static void AddEndpoints(this WebApplication app)
     {
-        app.MapItemEndpoints();
+        app.MapItemsEndpoints();
         // Add other endpoints here as needed
     }
 }
