+++ templates/MicroserviceTemplate.Aspire/tests/MyService.Tests/Items/EndpointTests/CreateItemEndpointTests.cs
@@ -5,9 +5,9 @@ using MyService.Domain.Entities;
 using System.Net;
 using System.Net.Http.Json;
 
-namespace MyService.Tests.ItemTests.EndpointTests;
+namespace MyService.Tests.Items.EndpointTests;
 
-public class CreateItemEndpointTests : ItemEndpointsTestsBase
+public class CreateItemEndpointTests : EndpointsTestsBase
 {
     public CreateItemEndpointTests(WebApplicationFactory<Program> factory) : base(factory) { }
